refactoring + all UTs pass
