+++ package.json
@@ -31,6 +31,7 @@
     "deep-freeze": "0.0.1",
     "eslint": "^9.18.0",
     "globals": "^15.14.0",
-    "memfs": "^4.17.0"
+    "memfs": "^4.17.0",
+    "vitest": "^3.0.4"
   }
 }
