--- src/lib/excel-file-manager.test.js
@@ -1,253 +1,0 @@
-import ExcelFileManager from './excel-file-manager';
-import { fs, vol } from 'memfs';
-import { jest } from '@jest/globals';
-
-// jest.unstable_mockModule('node:fs', () => fs);
-// jest.mock('fs', () => fs);
-
-beforeEach(() => {
-    vol.reset();
-});
-
-function createMockProcessor() {
-    let mockNum = 0;
-    return jest.fn(mockData => ({
-        data: mockData.data + 'DATA',
-        rowCount: ++mockNum,
-    }));
-}
-
-function createFileParser() {
-    const mocks = new Map([
-        ['excelblob1', {
-            mock1: {data: 'a'},
-            mock2: {data: 'b'},
-        }],
-        ['excelblob2', {
-            mock3: {data: 'c'},
-        }],
-        ['excelblob3', {
-            mock3: {data: 'c'},
-            mock4: {data: 'd'},
-        }],
-    ]);
-    return jest.fn(blob => mocks.get(blob));
-}
-
-function createInstance(overloads = null) {
-    const fileParser = createFileParser();
-    const mockProcessor = createMockProcessor();
-    const fm = new ExcelFileManager({
-        fs,
-        srcDir: '/mocks',
-        destDir: '/dest',
-        fileParser,
-        mockProcessor,
-        ...overloads,
-    });
-    return { fileParser, mockProcessor, fm };
-}
-
-test('should process dir (happy path)', async () => {
-    vol.fromJSON({
-        '/mocks/source1.xlsx': 'excelblob1',
-        '/mocks/source2.xlsx': 'excelblob2',
-        '/dest': null, // dir
-    });
-    const { fileParser, mockProcessor, fm } = createInstance();
-
-    await fm.processDir();
-
-    expect(fileParser.mock.calls).toEqual([
-        [ 'excelblob1' ],
-        [ 'excelblob2' ],
-    ]);
-    expect(mockProcessor.mock.calls).toEqual([
-        [ {data: 'a'} ],
-        [ {data: 'b'} ],
-        [ {data: 'c'} ],
-    ]);
-    expect(fm.mockList).toEqual([
-        'source1/mock1.txt',
-        'source1/mock2.txt',
-        'source2/mock3.txt',
-    ]);
-    expect([...fm.fileHashMap.entries()]).toEqual([
-        ['source1.xlsx', null],
-        ['source2.xlsx', null],
-    ]);
-    expect(vol.toJSON()).toEqual({
-        '/mocks/source1.xlsx': 'excelblob1',
-        '/mocks/source2.xlsx': 'excelblob2',
-        '/dest/source1/mock1.txt': 'aDATA',
-        '/dest/source1/mock2.txt': 'bDATA',
-        '/dest/source2/mock3.txt': 'cDATA',
-    });
-});
-
-test('should process dir with Hashing', async () => {
-    vol.fromJSON({
-        '/mocks/source1.xlsx': 'excelblob1',
-        '/mocks/source2.xlsx': 'excelblob2',
-        '/dest': null, // dir
-    });
-    const { fileParser, mockProcessor, fm } = createInstance({ withHashing: true });
-
-    await fm.processDir();
-
-    expect(fileParser.mock.calls).toEqual([
-        [ 'excelblob1' ],
-        [ 'excelblob2' ],
-    ]);
-    expect(mockProcessor.mock.calls).toEqual([
-        [ {data: 'a'} ],
-        [ {data: 'b'} ],
-        [ {data: 'c'} ],
-    ]);
-    expect(fm.mockList).toEqual([
-        'source1/mock1.txt',
-        'source1/mock2.txt',
-        'source2/mock3.txt',
-    ]);
-    expect([...fm.fileHashMap.entries()]).toEqual([
-        ['source1.xlsx', 'a2a3bb940141a764e09f92fa0011063ae4ed0d31'],
-        ['source2.xlsx', '4ec3b91ac2b025e71d0e870ce8f3d39e09bcb542'],
-    ]);
-    expect(vol.toJSON()).toEqual({
-        '/mocks/source1.xlsx': 'excelblob1',
-        '/mocks/source2.xlsx': 'excelblob2',
-        '/dest/source1/mock1.txt': 'aDATA',
-        '/dest/source1/mock2.txt': 'bDATA',
-        '/dest/source2/mock3.txt': 'cDATA',
-    });
-});
-
-test('should emit proper messages', async () => {
-    vol.fromJSON({
-        '/mocks/source1.xlsx': 'excelblob1',
-        '/mocks/source2.xlsx': 'excelblob2',
-        '/dest': null, // dir
-    });
-    const { fm } = createInstance();
-
-    const onStart = jest.fn();
-    const onMock = jest.fn();
-    fm.on('start-of-file-processing', onStart);
-    fm.on('mock-processed', onMock);
-    await fm.processDir();
-
-    expect(onStart.mock.calls).toEqual([
-        [{ name: 'source1.xlsx'}],
-        [{ name: 'source2.xlsx'}],
-    ]);
-    expect(onMock.mock.calls).toEqual([
-        [{ name: 'mock1.txt', rowCount: 1 }],
-        [{ name: 'mock2.txt', rowCount: 2 }],
-        [{ name: 'mock3.txt', rowCount: 3 }],
-    ]);
-});
-
-test('should process one file', async () => {
-    vol.fromJSON({
-        '/mocks/source1.xlsx': 'excelblob1',
-        '/mocks/source2.xlsx': 'excelblob2',
-        '/dest': null, // dir
-    });
-    const { fileParser, mockProcessor, fm } = createInstance();
-
-    await fm.processDir();
-    vol.writeFileSync('/mocks/source2.xlsx', 'excelblob3');
-    fileParser.mockClear();
-    mockProcessor.mockClear();
-    await fm.processOneFile('/mocks/source2.xlsx');
-
-    expect(fileParser.mock.calls).toEqual([
-        [ 'excelblob3' ],
-    ]);
-    expect(mockProcessor.mock.calls).toEqual([
-        [ {data: 'c'} ],
-        [ {data: 'd'} ],
-    ]);
-    expect(fm.mockList).toEqual([
-        'source1/mock1.txt',
-        'source1/mock2.txt',
-        'source2/mock3.txt',
-        'source2/mock4.txt',
-    ]);
-    expect([...fm.fileHashMap.entries()]).toEqual([
-        ['source1.xlsx', null],
-        ['source2.xlsx', null],
-    ]);
-    expect(vol.toJSON()).toEqual({
-        '/mocks/source1.xlsx': 'excelblob1',
-        '/mocks/source2.xlsx': 'excelblob3',
-        '/dest/source1/mock1.txt': 'aDATA',
-        '/dest/source1/mock2.txt': 'bDATA',
-        '/dest/source2/mock3.txt': 'cDATA',
-        '/dest/source2/mock4.txt': 'dDATA',
-    });
-});
-
-test('should process one file with Hashing', async () => {
-    vol.fromJSON({
-        '/mocks/source1.xlsx': 'excelblob1',
-        '/mocks/source2.xlsx': 'excelblob2',
-        '/dest': null, // dir
-    });
-    const { fileParser, mockProcessor, fm } = createInstance({ withHashing: true });
-
-    await fm.processDir();
-    vol.writeFileSync('/mocks/source2.xlsx', 'excelblob3');
-    fileParser.mockClear();
-    mockProcessor.mockClear();
-    await fm.processOneFile('/mocks/source2.xlsx');
-
-    expect(fileParser.mock.calls).toEqual([
-        [ 'excelblob3' ],
-    ]);
-    expect(mockProcessor.mock.calls).toEqual([
-        [ {data: 'c'} ],
-        [ {data: 'd'} ],
-    ]);
-    expect(fm.mockList).toEqual([
-        'source1/mock1.txt',
-        'source1/mock2.txt',
-        'source2/mock3.txt',
-        'source2/mock4.txt',
-    ]);
-    expect([...fm.fileHashMap.entries()]).toEqual([
-        ['source1.xlsx', 'a2a3bb940141a764e09f92fa0011063ae4ed0d31'],
-        ['source2.xlsx', 'c0f471714bec4543c45cfa75955b0a765a6b4f89'],
-    ]);
-    expect(vol.toJSON()).toEqual({
-        '/mocks/source1.xlsx': 'excelblob1',
-        '/mocks/source2.xlsx': 'excelblob3',
-        '/dest/source1/mock1.txt': 'aDATA',
-        '/dest/source1/mock2.txt': 'bDATA',
-        '/dest/source2/mock3.txt': 'cDATA',
-        '/dest/source2/mock4.txt': 'dDATA',
-    });
-});
-
-test('should fail on file from another dir', async () => {
-    vol.fromJSON({
-        '/mocks/source1.xlsx': 'excelblob1',
-        '/mocks/source2.xlsx': 'excelblob2',
-        '/another/source3.xlsx': 'excelblob2',
-        '/dest': null, // dir
-    });
-    const { fm } = createInstance();
-
-    // await fm.processDir();
-    await expect(fm.processOneFile('/another/source3.xlsx')).rejects.toThrowError('from another directory');
-});
-
-test('should fail on wrong params', () => {
-    const initiationFn = () => createInstance();
-
-    expect(initiationFn).toThrowError('Source dir');
-    vol.mkdirSync('/mocks');
-    expect(initiationFn).toThrowError('Destination dir');
-    vol.mkdirSync('/dest');
-    expect(initiationFn).not.toThrowError();
-});
