+++ vite.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+  test: {
+    environment: 'node',
+    exclude: [
+      '**/node_modules/**',
+      '**/_build/**',
+    ]
+  },
+});
